QueryResultRecorder: add support for ask, describe and construct
Queries seen in the logs might be of various types and sadly their
response format is not necessarily compatible with the binary tuple
format we expect.
For ASK queries we construct an artificial result set with a single
static binding and a single line holding the boolean.
For construct and describe we rely on the fact that blazegraph does
produce a json response that is parseable as a tuple query response to
obtain the tuples we expect (hence the new sesame rio dependency
required).

Some code has been ported from blazegraph itself to properly detect the
type of the query, main "difficulty" is to ignore the query prolog
(prefix definitions/comments) to obtain the first token that identifies
the query type.

Bug: T351819
Change-Id: If7f390efdbf3a41c710ae2b73cd63082c4cc8719