Use a simple expression syntax for defining throttling headers
We want to be a little more flexible defining when we throttle,
in particular the goal is to enable throttling when one header is
present and another header is not present. To make this possible
define an expression language with 2 operators, && and !, that
allow us to define this expression a bit more dynamically.

Bug: T361950
Change-Id: Ied1594dfb4f648c395bc720c2198752f77e33b4a