QueriesProcessor: Prevent bad optimizations
Poking at the physical plan of the dataframe here, spark decided
it was going to do `isnotnull(UDF(query))` before the
repartitioning, and then `UDF(query) as q_info` after the
repartitioning. Running this expensive UDF twice is a terrible
idea, especially because it runs it before the repartitioning can
spread the work out.

Resolve by marking the UDF as non-deterministic. That's a lie,
but it ensures that spark will not evaluate it twice and expect
the same result both times.

Suspect this will also solve the OOM problems, but will have to
run it a few times to see.

Bug: T371217
Change-Id: I26bf67f92e20715ab1d09c2031fd3794a0f66e08