Allow setting client side idle timeout
Without any timeout the client will keep the connection opened for ever.
The server might not be willing to keep forever and seems to fail after
a 30m idle confusing the client that is failing with an eof exception.
Set a 5min that should prevent such problems, it can be overridden by
setting the property http.proxyIdleTimeout.

Bug: T373812
Change-Id: Ic5bd8a7be793b16c533d137df83d4cbf51427b56