Add support for setting a start timestamp
When starting the application with a state that does not have the
offsets for a particular or an empty it might be handy to set the
initial offsets based on a timestamp.
Previous approaches involved using a python script subscribing to these
topic and resetting the offsets, this is particularly error prone and
dangerous.

Bug: T374919
Change-Id: I6f4d3277e096ab6fb099249ea4d9137379fedb1b