Add wdqs-subgraph-definitions-v2 with support for P13046
The introduction of P13046 (publication type of scholarly article) might
help to identify scholarly articles.
Add a wdqs-subgraph-definitions-v2 based on the v1 but with a new rule
based on the presence of statement using P13046.

Switch the spark job to use the v2, the updater job will have to have
its config updated to point to this new rule.

Bug: T376598
Change-Id: I04f9c883938819a893bff2d811ca423d5959c632