Add support for versioned streams
If use_event_streams_api is true the version ".v2" is appended to the
configured output streams.
For now a single version is supported but should give the opportunity
to build a separate set of kafka producer with varying serializers if we
ever need to deploy a breaking change with a ".v3" stream.

The version suffix is inferrend from the schema version provided via
the config:
- v1 (now deprecated) should force the use of the legacy mode that does
  not use eventutilities
- v2 and if the use_event_streams_api config option is set to true will
  suffix the stream names with that version

Also move some hard-coded constants (schema title & version) out of the
function body into dedicated constants in classes and/or interfaces.

Bug: T374919
Change-Id: Iacaf1c0a909c8fb21a4051ea78b3667e9dc48eac