streaming-updater-producer: use long for sequence fields
While json does not actually make a distinction between long & int the
eventutilities Row -> Json serializer actually cared and failed with
convertion issues.
Force the use of a long type for sequence and sequence_length fields and
also improve tests so that they should capture such problems in the
future.

Bug: T374919
Change-Id: I3d31c1c359958328b4b49a1edc59807b8f564ca5