Improve RDF dump parsing by not hard-coding prefixes
This list of hard-coded prefixes is fragile and assume that prefixes
are stable. These are not per RDF standards and should be extracted
from the dump.
Do a first pass on the first 1000 lines which we assume to contain the
prefixes, parse them and use them when reading chunks.

Relates to T384344

Change-Id: I94d16a1f35fc8e94fd63f44b2cc9464e1b45306c